Levels uses N of M like Stages; added missing space between N and "of"
One JS unit test is failing, but it already was before I changed anything.

Note that both Level and Stage still have names and still use the name as the value before "of".
Nothing yet makes Levels except the demo code. When we do it for real, we will make the names by auto-numbering as we do for stages.