Refactor thumbnail creation and fix it on Linux
We now use a background thread instead of the idle loop for thumbnail
creation. The GeckoWebBrowser functionality gets called on the UI thread by
Invoke calls.

This change also fixes the creation of thumbnails on Linux where the
ImageCreator functionality is not implemented. Instead we can use the GetBitmap
method of the OffScreenGeckoWebBrowser.