Add templates for Decodable Reader and Leveled Reader
These templates are mostly copies of the BasicBook templates, with their own guids.
Each adds to meta.json properties that cause the appropriate decodable or leveled reader tools to be enabled and shown.
The decodable reader also includes an additional stylesheet to make text larger and lines more widely spaced.
The htm files are simply generated from the jade and should not need independent review.
Basic Book.css is identical in both cases to the one from the Basic Book template, as are Factory-XMatter.css.