Handle use of either ICU 4.8 or 5.2 as needed
On Linux, we use the native ICU DLLs provided by the system. However,
newer versions of Ubuntu use version 5.2 and older versions use 4.8.
Because ICU embeds the version number in the API names, it's necessary
to use a different build of icu.net with the different native DLLs.

The correct version of icu.net is selected in a BeforeBuild step by
calling the icu-config utility to determine the version, and then
copying the right assembly into place from a subdirectory.

The logic is copied from similar code written for WeSay by
Chris Hubbard.