Download order from BloomLibrary now expected to include title
This allows title to have characters not allowed in the path of the bookorder part of the URL,
which allows us to put the correct title in progress dialogs.
The change also cleans up a few other places where special characters in title were not handled well.
Especially suspicious characters for further testing would be &<>=/\