Comment out Chorus initialization call (BL-414/BL-370)
On Linux we don't install the mercurial files as part of the Bloom installation.
Since we don't have S/R enabled for this release (BL-370) there is no point in
showing a warning message in release builds.

This change gets rid of the warning message and also disables the Chorus
initialization call - the latter for the reason to make it more noticable that
something needs to be changed when we enable S/R again.