Don't try to update files if target directory is read-only (BL-372)
If the files get different time stamps during the installation (because the time
wraps to a new second or minute) Bloom thinks that it has to update the files. If
the target file is in a read-only location, i.e. in the location where Bloom got
installed, this will fail. This change modifies the behaviour of the
UpdateSupportFile method so that it doesn't check for files needing to be updated
if the target directory is read-only which means we can't write the file anyways.