Fix BL-758 Make a way to remember window size
By default, we still want to just maximize each time. But Susanna,
who is doing videos, needs a way for Bloom to go to the same size
each time. With this PR, there is a new setting in user.config,
"MaximizeWindow". To enable this new feature, "MaximizeWindow" must be
set to "False" in a text editor. After that, the window size and
location will be remembered, using an existing class from libpalaso.