Fix BL-775, where the format button is in the wrong location
This was most noticeable on the cover, but showed up in other situations to. Previously, the format button was places as a sibling to the edit boxes; this would look ok so long as there was only one edit box in the translationGroup, but as soon as you had two or more languages, it wouldn't have any way of being located in the one with the current insertion point. With this commit, it is placed as a child of the edit box itself. This introduces a complication that the user can do a CTRL+A and delete it, but that seems a reasonable price to pay. If they need it back, clicking out and back in will re-create it.