BL-642, Don't use <base> in temp files
The goal here is to use smarts in our
local server instead of <base> to find
images that the HTML refers to
just using filenames (without paths).
Works for the full html document
because it is in the same folder.
The core change is in
HtmlDom.BaseForRelativePaths and in
EnhancedImageServer.MakeSimulatedPageFileInBookFolder()
(see comment there).
The temp 'file' now only exists in the
local server memory, but the url
makes it seem to be in the book folder.

Since we're not using base, all temp
files that need to pretend to be in
the book folder must now use the
local server. This resulted in
changes to how we handle cases
where we want to use the full-resolution
images, like PDF generation.