BL-785: prefer to find xMatter files in Factory XMatter
This is the original source for files that are copied to other
places in individual templates that are also distributed
with the application. We want live editing to use the most
original version so we see changes when editing that.

New tests and other changes try to document and
preserve other orderings we care about.