Fix BL-786: cursor half out of the box, typing does nothing
The problem is a regression; we previously had a workaround that was
causing a different problem and was recently removed. This time around
we insert an empty <span> to give gecko some bearings so it doesn't
focus on our language qtip. We then remove the span when leaving the
box.