Fix BL-777, where the languagetip was showing in the wrong edit box.
A previous workaround for 2 gecko bugs worked like this: If the focus
is in the text box, hide :after tag that was showing the language name,
and instead add a separate div inside the translation group. But in
doing so, it had no way of being in the right place
for bi and tri-lingual modes.

Due to a previous fix (PR#173), this is all unneeded now, and this
commit just takes us back to just relying on the css-positioned
language tag.