Disable checking for updates on Linux
On Linux Bloom gets installed through the package manager which
regularly checks for new versions. Having the application check
for new versions instead will not work because the user doesn't
have permission to replace the existing executable, and the current
code would replace it with the Windows instead of the Linux version.
Seems better to leave the update job to the system package manager.