Fix BL-799: hovering on format cog makes the source bubble disappear
For now we'll just not have the tooltip on the cog. It was also in the way once the format dialog came up. In the future, we could move it to qtip2, so that it wouldn't interfere with the qtip being used by the source language bubble, and change somebody's zindex so it doesn't interfere with the format dialog.