Don't cut off name when creating source collection (BL-733)
Cut off was caused by BetterLabel which was to high, overlapping
the text box. We don't really need a BetterLabel here, so I
replaced it with a regular Label control.

Also anchored the text box and the HTML label that shows the path
to the right so that the full width of the window gets used.