Fix failing unit tests on Linux builds on Jenkins
The unit tests that access S3 fail on Jenkins because the build
agent is behind a proxy.The Amazon SDK classes don't make use of the
http_proxy environment variables, causing the tests to fail.

This change retrieves the proxy settings from the http_proxy or
HTTP_PROXY environment variables and sets the corresponding
properties for the S3 client.