BL-830,647: more complete upgrade of jquery-ui
This fixes the original BL-647: format dialog can
now be dragged even if page is scrolled.
The previous fix broke decodable readers
- needed current jquery-ui.css to give right z-index
to background and dialog
- then found we needed an appropriate theme to
make accordion somewhat match rest of program.
I haven't yet figured out a theme that exactly matches
the old one, but this is similar enough to work,
though perhaps not as aesthetic.
- Also the system of attaching ids and controlling
which tab is active in the accordion has changed,
had to redo that.