Add code to help track and maybe help the problem in BL-836
BL-836 is where at times the reader is getting an object = null. We can tell from the line number that it is either CurrentBook or CurrentBook.CollectionSettings. Possibly this is a timing problem. This commit doesn't attempt to get at the root problem, but catches it earlier with a debug message, and returns false,