Allow multiple test runs at the same time
Now that we have the tests running on both Windows and Linux some
of the tests randomly fail. It looks like it's caused by a race
condition when the same test runs multiple times at the same time.
This change improves several of the tests that interact with
Amazon S3 by adding a random part to the username and directory
name.