Allow user styles to effect divs that are not editable
I'm not clear why, but previously the StyleEditor code assumed and required that user modified styles began with .bloom-editable. This made it impossible to have make a page where, if the user modified the style of some editable box, other elements on the page would change as well, to keep the page consistent. The particular page for which this came up was the credits page, where a big chunk of the page is not .bloom-editable's.