Make S3 tests more robust
Our previous implementation of the S3ClientTests had the problem
that at the beginning of each test we emptied the test bucket.
This can cause test failures if the unit tests run at the same
time on multiple machines.

The new implementation only deletes the books the test uploaded.
This allows multiple tests to run in parallel, but has the
disadvantage that we might leave behind some test data if the
developer stops a debug session while debugging the unit tests.
However, I think this will happen rarely and it won't cause other
tests to fail.