BL-862 Error accessing downloaded book in Linux
When a book downloaded from BloomLibrary is accessed,
either by the attempt to open immediately after downloading
or by clicking on the book, an exception was thrown saying
it was unable to find the ../customCollectionStyles.css file.
This was because the href ..\customCollectionStyles.css in
the book html needed to be normalized prior to using it to
locate the file.