BL-822 Text entry is not usable
Very similar to WS-62 with a very similar change and
similar comments. On Linux systems, with specific fonts
such as Arial 12pt, entry into a text box displays with
characters oddly compressed and occasionally invisible.
This change corrects the display of those characters.

The JIRA  also reports issues with the cursor position.
Once displayed properly, the issue is somewhat less
apparent, but can still be seen towards the end of the
line. This change does not address that part of the
problem.  Changing the font by a point can be a workaround.