Fix BL-886: Border control defaults to solid, when it should be "none"
The odd thing is that the code looks at the box on screen, rather than the rule this dialog is supposed to be editing. So the Edit Mode borders were interfering. The code assumed such borders were less than 50% opaque, but actually the selected box one is 60%. Changed to detect anything less that 100%.