Fix BL-889: Crash when entering Edit for a second time on a new book
When we leave the edit tab of a newly created book, bloom actually reloads the book from disk. At this point, the book creates its page cache. Later, when we go back to the edit tab, it rebuilds its XMatter, which means removing and re-adding pages. At that point, some of the pages in the cache have ids to xmatter pages that have been replaced. So a number of approaches to fixing this present themselves, but I went with invalidating the cache when we know its bad.