Fix BL-912: immediate title page response to change of settings
Previously, you had to change pages and come back before the location would be updated. So the first step was to make a PageRefresh event that the settings could call. However, this lead to another problem, which is that the old location text would get sucked into the dataset, and then pushed back out. It was overriding the newly updated "location" value. It overrode it because the computed location value is currently in "neutral" language, "*", but on the page it goes into a box that asks for the national language. If there is none in the language, fine, the "*" one is copied in. But during refresh, the national language one gets sucked in, so now, as the new page is constructed, that sucked in (old) version of "location" wins out over the generic "*" version.

A otherwise reasonable fix would have been to change the template to specify lang="*". However I'm reluctant to do that because the lang is also used to choose the font. I took the safer route of introducing a "bloom-writeOnly" class on this element, so it never gets sucked in.