Fix BL-904 Format dlg's width should not be limited to that of...
Fix BL-904 Format dlg's width should not be limited to that of the parent text box. I don't see any reason for the code to limit this width, so I removed it. Perhaps the reason will present itself at some time in the future and then we can handle it and have a code comment this time.