Fix BL-875 format dialog tooltips make Source Bubbles disappear
QTip, at least our version or the way we use it, can only have one tooltip on an object. Apparently this also prevents having a tooltip on the parent translation box and a tooltip in the format box, which is a child. Fortunately we already have a "qtipSecondary" so I switched the format dialog to use that exclusively.