Work around Bl-923: L10NSharp Crash when accessed from javascript
I could not reproduce the bug immediately after it happened to me, even after removing my localization files. Looking at the l10nsharp code, the only way I can see us getting this error ("Collection was modified; enumeration operation may not execute.") is by re-entering the code on a different thread.

We could make the Bloom server be thread-safe, but we could also do that to L10nsharp.

Meanwhile, this patch catches the error and swallows it for users.

Needs review: previously, the code got the translation, then decided whether it needed the translation or not. Now it first decides if it needs it. Is there something I'm missing about the original approach?