Fix dialog message display (BL-871)
The original bug was widget overlap clipping some text characters.
When that was fixed, the message was breaking lines in the middle of
words.  I fixed that by changing from a Label to a TextBox for the
message display.  This is very common in our code for multiline text
displays.