Work around BL-938: Crash after changing UI Language
This happens under difficult to reproduce situations, where some combination of changing UI language and opening a new collection leads to L10NSharp trying to change the text of the "Duplicate This Page" tooltip and then, farther into .net, getting a null reference exception. Assumption is that something (L10NSharp?) is holding onto the control that has been garbage collected, so doesn't appreciate being woken up to get a new label.

For now, this just detects the situation and swallows it in release mode.