Fix BL-933 If text box is empty, typing disables the format button
And so we continue the search for the best work around to the FF cursor problem :-) In this episode, we extend the selection to include the &nbsp; we want to delete, just before the first typing or pasting happens. (the &nbsp; is in there because without it FF doesn't know where to put the cursor, since we have an ":after" pseudo element).