Allow empty lines at the start.
This comments out a fix to a lesser problem, which was that when we do a CTRL+A DEL, FF leaves us with a <br></br> at the start. When the first key is then pressed, a blank line is shown and the letter pressed shows up after that.

So currently that bug is back. Perhaps in the future we can act when the DEL or Backspace occurs and then detect this situation and clean it up.