BL-928 Restrict Json uploaded to parse.com
Prevent uploading a field that was recently added to BookMetaData
but which is not relevant to our website, and thus allow uploads
to work again even though Parse.com does not find the field.
Provides a general way to contro which fields get uploaded;
in future, new metadata fields should only be uploaded if actually
useful to the web site.