Allow Book tests to run with newer Moq versions
Moq 4.3 (Moq >= 4.2.1408.x) introduced the automatic mocking of
interfaces. This has the side effect that the c'tor on
HtmlThumbNailer gets called but no longer the Dispose() method,
leaving a thumbnailer hanging around. This causes subsequent tests
that try to create a thumbnailer to fail.

This change modifies the code to directly use a HtmlThumbNailer
object instead of a moq. This is acceptable since no other methods
besides the c'tor and Dispose() get called on that object.