Rename LowResImageCache to RuntimeImageProcessing
Really, that's all that was done here. It just affected a lot of files. This is being done in this commit so that in a follow-on commit, we can actually change what this class does. Changing the name in its own commit is to make it easier to see what changed, in the follow-on commit.