Do less with images, while being more thread safe.
Now that we deliver images over http, we get around the problem of gecko giving up on large files too easily. Now, we can deliver very large images (I tried up to 9mb jpeg) directly, at gecko does a good job of showing  low-quality img quickly and then enhancing it a it gets time. So, for jpegs, we now don't touch the image at all, except to read it and push it over the wire.

Now, for PNGs, we do still redraw them in order to introduce transparency. It it is conceivable that we should still shrink png images, but it's not clear if we should, so I'm going with no shrinking for now as the simpler operation, and then we can reintroduce it if testing shows it is needed. The code is still there #if'd out.