Fix BL981 & BL1028, source bubbles & data-hint bubbles covering text
This fix uses the following heuristic: if we know the text box we're pointing at uses up the full width of the marginbox, then it is safe, horizontally, to show it (it may still not be save vertically, this doesn't address that). Otherwise, error on the side of caution and only show the bubble when the user clicks in it.