BL-546 Decodable Reader Tool PUA character display issue
Decodable reader tool not displaying characters and words with
PUA characters.  If the user set the font for the current
languages to Charis SIL and entered PUA characters into
the Letters and Letters Combination, they displayed
correctly on the Stages Setup screen, but not on the
accordion screen displaying the letters and words for
the current stage.  The fix applies the user specified
font to the letters and words divs on the accordion screen
as they are applied on the setup screen.