Introduces CommandAvailabilityPublisher and websocket-based messaging
Commands have a status that needs to be communicated to the html client. The CommandAvailabilityPublisher does that by wiring up to events on commands and sending updated status on commands over a websocket. This uses the "fleck" nuget library to implement websockets on the server side (they are built-in to Firefox on the client side).
This doesn't do anything the user can see yet. It's a bit of infrastructure that we will build on later. In particular, nothing on the javascript side is yet listening for these messages. Also, more work needs to be done about automatically wiring up Commands; see the comments in ProjectContext, where the suggested approaches didn't work, so only the DeletePage command is currently wired up, by hand.