Fix BL-1060, Trying to save page in the wrong book
Though we could never reproduce this bug, it did get reported several times. The last time, I had the user's collection, and found that Book A was being asked to save a page that came from book B. I'm unable to see how this is possible, or to reproduce it. However I did add in a bunch more logging which might help when the situation happens again. But even when it happens, we won't have the same failure because now it gets which book should do the saving from the page itself, rather than the "current book".