BL-846 Format Dialog doesn't always close when another box gets focus
If you have the format dialog open in a text box and click in
another text box not currently in focus at exactly the point the
gear icon will appear in that box, focus shifts to the new box but
the format dialog for the original text box remains open.  The fix
makes sure that the focus event triggers the removal of any open
format dialogs as it shifts the gear icon from one box to the other.