Migrate old SHRP books that only had English titles.
Bloom has been changed to prefer "No Title" or somesuch if the vernacular title is missing. But we have hundreds of books with only English in the SIL-LEAD/SHRP project. Fortunately these books have a set pattern for names, so this commit detects the situation and allows the English name to be displayed.