BL-1014 Format Dialog text not translated
If you are using a non-English UI, the bottom line of text in
the format dialog will initially appear in English.  If you
press the format dialog for another field on the same page, the
localized text will appear.

The localization manager GetText function, which is called to
retrieve the text by the StyleEditor.ts file, returns English
the first time it is called unless the strings are already loaded
into the dictionary.  This change uses an asynchronous call to
get the text instead.