BL-1014 Format dialog text not translated
The other problem was that the BookEditor.DefaultForText field
was not present in the tmx file.  This also will need translations
for French before the entry will always consistently be French on
the characters tab, since half of the time it uses this string and
half of the time uses another string that has been translated.