BL-1010 Deleting resized split on custom page produces error
This also corrects BL-1043 which is essentially a duplicate. Creating
a split on a custom page, resizing it and then deleting it resulted in
the remaining split from the pair to be the wrong size.  This
occurred with either vertical or horizontal splits.  The fix provided
here saves the width and height of the original parent and applies it
to the updated sibling which is replacing it, as it currently saves the
class.