BL-788 [Trusty Linux] Print dialog opens behind Bloom window
The print dialog appears behind Bloom on Linux except for the
first time it is displayed.

This solution is a compromise since I was unable to find a way
to resolve the issue by bringing the print dialog to the foreground
in all cases.  This solution minimizes Bloom when the print button
is pressed on a Linux system.  The print dialog displays.  When the
user either presses cancel or print, Bloom is restored to the
foreground at the same size it was before Print was pressed.