Add logging and enhanced error handling plus refactoring/clarifying
The central thing here was changing how we handle problems encountered when we look for the version table. We now create a result-object that has the url if we got one, and and the exception if we got one. The exception is logged and the user is notified at the point that we get it, instead of leaving it to clients. But clients get the full object, and can continue to decide what to do, as before.

This also normalized code for determining the channel name.

Logging happens so that if there is a problem, the log will tell us what happened.