BL-1040 Show book name while editing
This change adds the bookname to the application title bar.
When the application starts or a book has been deleted so that
there is no selection, the title appears without a book name.
Otherwise, any time a book is selected, it appears first, followed
by the collection name, followed by the release and build
date information.