Fix BL-1121 by making a number of fields to be copied when empty
The problem appears to be related to fields that are in a language, like credits, but which we really don't want to not show at all if the shell is used for some other language. These kinds of fields can be marked with a special class, bloom-copyFromOtherLanguageIfNecessary. This commit applies that to a few more fields. So far so good. However the original implementation didn't take into account the case of fields that are vernacular, like the front cover credits. I don't feel like I'm in a position to revisit that at the moment, so I made an exception for that one field. See the comments in the code.