BL-831 Unit Test Hangs
Fixes the one unit test that was a true failure.  The others
are Jenkins only environment failures that work fine when
run on developers systems.  This one runs into problems
because the ToString isn't supported for a type "object" and
clearly states that it doesn't.  It also isn't called in the
code base that way.  This creates the same string another
equivalent way and verifies that the LocalData segment is empty.