Fix BL-1031 by saving as jpeg when a pasted image would be 50% smaller
BL-1031 is "Copied jpegs from Libre Office are delivered to Bloom as PNGs, which expands size dramatically". The problem this fixes is that libre office puts jpeg images onto the clipboard as bitmaps. So if the image is a complex color photo or painting, if we save as a PNG, we may dramatically increase the size of the image. This commit saves it both as a png and as a 100% quality jpeg, and compares the resulting file sizes. It chooses jpeg if that would be more than 50% smaller.