Factor out toolbox code from bloomEditing.js and remove template
The old "Custom Template" template is now obsolete. But it represents some work that we could well return to in a future version. So this commit moves the template to a location where it won't be seen by the user, and factors out the supporting code to its own file; that file is not loaded.