Fix book thumbnail tooltips to show full titles (BL-817)
This uses a previously unused ToolTip object.  AutoEllipsis was
causing truncated tooltips to be shown on Windows, but no tooltips
were displayed on Linux for the book thumbnail buttons.  It doesn't
appear to me that the purpose for AutoEllipsis described in comments
was valid, so I just disabled it with a comment after removing the
previous comments.