BL-1138 Book counts inaccurate after switch from DR
Leveled Reader Tool's This Book counts inaccurate when used from
Decodable.  When you create a new page with the decodable reader
tab visible, enter text, and then switch to the leveled reader tab,
the counts are wrong until another word is entered.  This is because
the doMarkup code adds in the current page information, but it
is not added in with the call from updateWholeBookCounts.