Fix BL-1242, when book is saved, span.bloom-linebreak gets removed
We already had code that did this on read, but not on write. So the "active ingredient" in this fix is simply to factor out the bit of workaround code we were using during reading so that it can be re-used during writing.
To facilitate the processing of the content, I switched the xmlwriter to work in-memory.
I could have refactored this even further, but it's at the very core of things and I'm just trying to get a priority fix out for Uganda.