In folios, rewrite the img src attrs so that the image is found
We recently switched from file-path to URL-based PDF creation. That worked fine for most books, because for those images are just a file name. However for folios, where many books are combined, we need a relative path to the folder where the image resides. So now when a folio is being assembled for running through the PDF engine, an img src attribute might be changed from "lion.png" to "../animals/lion.jpg", which then gets url encoded.