BL-2174 Make HtmlConverter save intermediate result in RAM
Stack trace in BL-2174 indicates that somewhere deep in libtidy
it fails to write a byte because a stream is mysteriously closed.
Could not figure out why, but there may be some issue with writing
the same file twice in close succession. It's better anyway not to
hit the disk with the intermediate result, and the change seems to
fix the problem.