BL-2399 Repair 'Report a Problem'
YouTrackSharp depends on EasyHttp.dll. The compiler
should realize this and 'copy local' both of them, but
doesn't, possibly because EasyHttp.dll is not at the
expected relative location. Adding a direct reference
makes the DLL available when needed.

The other change tweaks control sizes and positions
so the status messages can be fully read.