Fix BL-2501 Prefer ckeditor's undo if available. Else use document's instead of geckofx's.
This seems to fix the crash of BL-2501 and seems to give the desired undo behavior even of stuff ckeditor does (e.g. formatting).

While we're at it, it's normally easy enough to ask the browser to undo from within javascript, so I changed it to do that rather than leave it to c# to do that.