BL-2629 Generate thumbnails for Add Page dialog
The template folder may now contain a Caption.png file,
which will be used if Caption.svg is not found. If neither is
found, the system will generate a thumbnail on the fly.
Achieving this involved some tricky changes to our fileserver
so requests for multiple thumbnails don't lock out the server
requests needed to generate a thumbnail.