Fix BL-2738 by having source bubbles preemptively remove competitors
The problem was that we have the hint bubbles, e.g. "Book Title In English", showing up under or over the new Source Bubbles. It would be nice to have both, but we don't have the space.

Now, if a Source Bubble is going to show, e.g. on Book Title, it removes any label.bubble elements that are in the same translation group. If it isn't going to show, they stick around. Which is what we want.