Fix failing unit test on TC only by fixing test setup.
The test had not been set up to actually locate xmatter files in distfiles. The tests only succeeded on dev machines because they were able to find the needed files elsewhere. Presumably the agent also used to have some book around that was getting used, and someone deleted it.

So here, I not only fixed the problem, but introduced a test that can't be fooled because it introduces a new "ForUnitTests-xmatter" that won't be found in some other book.

This xmatter is screened out in CollectionSettingsDialog so the user won't see it.