Fix BL-2809 by not allowing booklet option when you choose letter
Limiting this actually helps shape the user's mental model towards how this really works; if they think "I'm going to print a booklet with my letter paper", then really we need them to choose "half letter" anyhow. Unlike with metric sizes which can be easily doubled/halved, double of "letter size" is... unusual. We can always enable this if there is demand for it.