Use Automatic Package Restore for nuget
This replaces the previously used MSBuild integrated package restore
and should enhance the build speed a little bit since it doesn't
restore the packages each time we build but instead when loading
the solution in VS/MD, or when building from the command line (using
Bloom.proj). It also allows to build when not connected to the Internet.