Fix BL-2847 epub does not show images that have spaces
It's not clear why the code was previously changing spaces to underscores, but it appeared to be done in the saving of the file, but not in the html, at least in the case reported in BL-2847, which was a cover page and so may have something special going on. As a potential fix for JT to review, I've just removed that conversion as I can't see a reason to do it, it could potentially give us the wrong image (if 2 images differed only by this distinction), and it seems to work fine without it.