Fix BL-2877 Losing cover page image
Previously I didn't expect the cleanup to run at the point where the book didn't yet have its xmatter applied. But I was wrong, and so this fix now also protects from deletion any filename that is mentioned in the datadiv, which is there even when the xmatter is not yet applied and thus not yet referencing the cover image.