Help BL-2871 by limiting the transparency operation to 1000x1000
1k x 1k is suitable for screens. It will not effect what we get in the PDF. Oddly, on the 6k x 6k "Opposites" book, it actually looks better than the original.

Also, changed the retry code to only wait a short while before giving up. The old code would wait 5 seconds. 5 seconds * 20 pages == "Bloom is dead".

A good follow-on operation would be to limit this transparency operation to only the cover page. Tricky, because the server has no idea, so we'd need to get a query onto the image path to specify that it is on cover stock.