BL-2980 remove spurious ref to notify-custom.js
This reference was added in commit
96e373863596ca056fd201fac7fdada0824b1027
on August 17 2015, apparently to support a single .notify
call that was added to page-chooser.ts for debugging
and accidentally not removed. The .notify call was commented
out two days later and sometime later removed completely.
The notify-custom.js file is also gone. I think we can
safely remove the code that tries (and fails) to load it.