Start using NPM for JS
This is just a baby-step... we still have to check in the required files because our build system would just ship it all (we don't want to check in all of node_modules, just the files we use). Also, we still have to add script elements for each one, manually. A future step would be to employ something like browserify so we can require/import files.