Starting to move code under toolbox into a single webpack entrypoint
This entails a tone of cleanup, as this set of tools is more strict, less friendly to global soup, etc. So it's worth doing just for that. Later it will make it feasible to use react and other modern javascript libraries that are painful without a module system.

At this point there are still countless issues.