Fix BL-3160  audio backup file occasionally locks stopping recording
This fix detects that and makes a new backup. I was tempted to try a simpler approach which would be to record to a randomly named temp file, then copy that to our destination only if we are satisfied. The current approach instead always makes this backup that we can revert to. I left it because if we add undo, this will be useful.