Fix BL-3153 where browser asked for audio file before it was written
On at least my machine, it was never possible to record something and play it back without switching pages. The reason is that at the end of the recording, the javascript would signal the html audio element that the source had changed. That would immediately hit the server up for the new file, but meanwhile the thread writing that file was still working on it. I tried various approaches but the cleanest ended up just telling the audio element not to preload the audio... just get it when the user clicks on "play".

In the process, I added some error handling around getting files, returning appropriate http codes.