Fix BL-3115, and a number of small probs when html or json is corrupt
The proximate problem was saving the book, in the constructor when it lacked a color and was too corrupt to be saved. We no longer save during construction. While I was in there, I made newly created books have the same cover color as the preview the user was shown.

Corrupt meta.json used to use a very different code path to talking to the user; changed that so that they get the same UI, with a button to send us the book.