Prevent book.userprefs getting all NULs
We have several unexplained cases of NULs in userprefs. Looking at the code, if the serialization failed, perhaps it would cause this.

So now we check to see that at least we have a real string. In case the real problem is shutting down while it's being written, we now do a separate safe write and then do a simple "replace" (using a palaso library class) which should be atomic.