Add new bloom-alignImageTopLeft class for inline images
For inline images with captions, we want the image container to grow as much as needed to fit the image and the caption (which can be multiple lines). But prior to this new class, all images were always centered within the image-container. That made it impractical to have an image-container without a fixed height. With this, we can add this class to turn the centering mechanism off.