Fix BL-3259 Can't set cover to an image with a '+' or '%' in the name
Turned out to be two separate problems.

The '+' one: This did not have a satisfying solution... in part because the rules for encoding are just too conflicting (or perhaps I'm missing something). Of course + can mean SPACE, and indeed we use it that way when sending url query string to the server. Howver our image urls have + as... +. But when UrlPathString.CreateFromUrlEncodedString decoded that, it got rid of the +, changed it to space.

So I made the code defeat that bit of HttpUtility.UrlDecode(), so that '+'s stay as '+'s. Added more unit tests, did a bunch of by-hand tests, and all seems good.

The '%' one: The solution for this is on the URL-encoding side. All explained in UrlPathString.UrlEncoded.