BL-3305, handle installation in Program files
Depends on https://github.com/BloomBooks/Squirrel.Windows/pull/17,
so merge that first.
When the installer is run with the new --allUsers command
line argument, it will install Bloom in Program Files (x86).
These changes mean that in this situation Bloom will make
all its registry entries in LocalMachine instead of CurrentUser.
Also, in this (typically classroom) situation we assume an admin
is managing updates, and Bloom will not be able to do its own in
Program Files, so we don't attempt automatic updates.