Allow people working in source collections to add books from bl.org
This is how it was for most of Bloom's life. At some point I decided to disallow it in order to enforce the distinction between collection types, but it turns out to almost be a blocker for people doing legitimate things like making a bloompack from books on bl.org. We should do more thinking on this, but I remembered I promised to remove this restriction while we think it over.