Move templates and sample shells into bloombrowserui
The goal here was to bring these less and jade files into the build system and stop committing build artifacts (css and html). This required three things: 1) we had to allow "html" in addition to "htm" so that we  can just use standard jade and 2) the 'attributes' change so that we can use modern jade and 3) change where code looks for these files.