BL-3438 Show toast when saving page
Previously, the websocket was only used by the audio system. This changes it to be more general, and then makes the page list view show a localized notification when pages are being saved. You won't see the message if it's be saved while leaving this tab or quiting the program, that's ok.